requirements.txt
streamlit
tavily-python
beautifulsoup4
requests
google-generativeai
langchain
langchain-core
langchain-community
langchain-google-genai
python-dotenv
fpdf2